Orient files around changesets
