+++ deployment/ccip/changeset/deploy_chain_test.go
@@ -1,6 +1,8 @@
 package changeset
 
 import (
+	"encoding/json"
+	"fmt"
 	"math/big"
 	"testing"
 
@@ -90,4 +92,20 @@ func TestDeployChainContractsChangeset(t *testing.T) {
 		require.NotNil(t, state.Chains[sel].OffRamp)
 		require.NotNil(t, state.Chains[sel].OnRamp)
 	}
+}
+
+func TestDeployCCIPContracts(t *testing.T) {
+	lggr := logger.TestLogger(t)
+	e := NewMemoryEnvironmentWithJobsAndContracts(t, lggr, 2, 4, nil)
+	// Deploy all the CCIP contracts.
+	state, err := LoadOnchainState(e.Env)
+	require.NoError(t, err)
+	snap, err := state.View(e.Env.AllChainSelectors())
+	require.NoError(t, err)
+
+	// Assert expect every deployed address to be in the address book.
+	// TODO (CCIP-3047): Add the rest of CCIPv2 representation
+	b, err := json.MarshalIndent(snap, "", "	")
+	require.NoError(t, err)
+	fmt.Println(string(b))
 }
