+++ deployment/ccip/changeset/token_info.go
@@ -11,6 +11,18 @@ import (
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 )
 
+type TokenSymbol string
+
+const (
+	LinkSymbol   TokenSymbol = "LINK"
+	WethSymbol   TokenSymbol = "WETH"
+	USDCSymbol   TokenSymbol = "USDC"
+	USDCName     string      = "USD Coin"
+	LinkDecimals             = 18
+	WethDecimals             = 18
+	UsdcDecimals             = 6
+)
+
 var (
 	TestDeviationPPB = ccipocr3.NewBigIntFromInt64(1e9)
 )
